Dissolve container in component
