+++ src/components/BudgetItemFormComponent.jsx
@@ -1,4 +1,6 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 import {
   Button,
@@ -13,7 +15,14 @@ import {
   Typography,
 } from '@material-ui/core';
 import AddIcon from '@material-ui/icons/Add';
-import BudgetGroupForm from '../container/BudgetGroupFormContainer';
+import {
+  getCurrency,
+} from '../redux/modules/AppReducer';
+import {
+  actions,
+  getIsBudgetGroupFormOpen, getBudgetGroups,
+} from '../redux/modules/BudgetReducer';
+import BudgetGroupForm from './BudgetGroupFormComponent';
 
 class BudgetItemFormComponent extends Component {
   state = {
@@ -158,4 +167,15 @@ BudgetItemFormComponent.propTypes = {
   currency: PropTypes.string.isRequired,
 };
 
-export default BudgetItemFormComponent;
+const mapStateToProps = state => ({
+  open: getIsBudgetGroupFormOpen(state),
+  budgetGroups: getBudgetGroups(state),
+  currency: getCurrency(state),
+});
+
+const mapDispatchToProps = dispatch => bindActionCreators(actions, dispatch);
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(BudgetItemFormComponent);
