--- src/container/BudgetSummaryContainer.jsx
@@ -1,21 +1,0 @@
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-
-import {
-  getCurrency,
-} from '../redux/modules/AppReducer';
-import {
-  actions,
-} from '../redux/modules/BudgetReducer';
-import BudgetSummaryComponent from '../components/BudgetSummaryComponent';
-
-const mapStateToProps = state => ({
-  currency: getCurrency(state),
-});
-
-const mapDispatchToProps = dispatch => bindActionCreators(actions, dispatch);
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(BudgetSummaryComponent);
