--- src/container/IncomeDeductionsItemContainer.jsx
@@ -1,21 +1,0 @@
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-
-import {
-  getCurrency,
-} from '../redux/modules/AppReducer';
-import {
-  actions,
-} from '../redux/modules/IncomeReducer';
-import IncomeDeductionsItemComponent from '../components/IncomeDeductionsItemComponent';
-
-const mapStateToProps = state => ({
-  currency: getCurrency(state),
-});
-
-const mapDispatchToProps = dispatch => bindActionCreators(actions, dispatch);
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(IncomeDeductionsItemComponent);
